requirements.txt
# 科研论文专用图片组合工具依赖包
# Python版本要求: >=3.8

# 图像处理核心库
Pillow>=10.0.0
numpy>=1.24.0

# GUI界面库
PyQt5>=5.15.9

# 矢量图支持（可选，用于EPS格式）
matplotlib>=3.7.0

# PDF/EPS处理
reportlab>=4.0.0